kata: clean up the init project, moving model around and creating my first singnalr hub;
