+++ SignalRDemo.HubApp/Hubs/DataHub.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.SignalR;
+
+namespace SignalRDemo.HubApp.Hubs;
+
+public class DataHub : Hub
+{
+	public async Task BroadcastData(string data)
+	{
+		await Clients.All.SendAsync("From DataHub, ReceivedData:", data);
+	}
+}
